requirements.txt
PyMySQL~=1.0.2
Django~=3.2.10
requests~=2.28.1
qrcode~=7.3.1
setuptools~=65.5.0
Js2Py~=0.71
Pillow~=9.2.0
